backend/requirements.txt
asgiref==3.2.10
Django==2.2.10
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary
